fixed stack position detection for powerpc
git-svn-id: svn+ssh://svn.code.sf.net/p/qore/code/qore/trunk@2069 1b0b35fe-f740-4c1b-b404-a338b32e513b

+++ lib/thread.cc
@@ -343,7 +343,7 @@ class ThreadData
       ClosureRuntimeEnvironment *closure_rt_env;
 
 #ifdef QORE_MANAGE_STACK
-      long stack_limit;
+      size_t stack_limit;
 #endif
 
       DLLLOCAL ThreadData(int ptid, QoreProgram *p) :
